refactor env_provider into settings module, keep DAG configs separate
